q1/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
python-multipart==0.0.9
markdown==3.5.2 